Refactor and put tests in namespace
